+++ R/build_livestock_nutrient_flows.R
@@ -34,14 +34,14 @@
 #' intake <- tibble::tribble(
 #'   ~year, ~territory, ~sub_territory, ~livestock_category,
 #'   ~item_cbs_code, ~feed_quality, ~intake_dm_t,
-#'   2020L, "ESP", NA, "Cattle_milk", 2513L, "high_quality", 200,
-#'   2020L, "ESP", NA, "Cattle_milk", NA, "grass", 600
+#'   2020L, "203", NA, "Cattle_milk", 2513L, "high_quality", 200,
+#'   2020L, "203", NA, "Cattle_milk", NA, "grass", 600
 #' )
 #' gridded <- list(
 #'   crops = tibble::tribble(
 #'     ~year, ~territory, ~sub_territory, ~crop, ~manure_n_receptivity, ~crop_n_cap,
-#'     2020L, "ESP", NA, "barley", 6, 200,
-#'     2020L, "ESP", NA, "wheat", 4, 200
+#'     2020L, "203", NA, "barley", 6, 200,
+#'     2020L, "203", NA, "wheat", 4, 200
 #'   )
 #' )
 #' build_livestock_nutrient_flows(intake, gridded = gridded)